requirements.txt
discord.py==1.6.0
sqlalchemy==1.3.23